refactor: extract `enabledMaterials` to BlockMaterialPipe class
