+++ src/main/java/gregtech/common/pipelike/itempipe/BlockItemPipe.java
@@ -4,7 +4,6 @@ import gregtech.api.items.toolitem.ToolClasses;
 import gregtech.api.pipenet.block.material.BlockMaterialPipe;
 import gregtech.api.pipenet.tile.IPipeTile;
 import gregtech.api.pipenet.tile.TileEntityPipeBase;
-import gregtech.api.unification.material.Material;
 import gregtech.api.unification.material.properties.ItemPipeProperties;
 import gregtech.api.unification.material.registry.MaterialRegistry;
 import gregtech.client.renderer.pipe.ItemPipeRenderer;
@@ -16,44 +15,26 @@ import gregtech.common.pipelike.itempipe.tile.TileEntityItemPipeTickable;
 
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
-import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumBlockRenderType;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.NonNullList;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import net.minecraftforge.items.CapabilityItemHandler;
 
-import com.google.common.base.Preconditions;
 import org.apache.commons.lang3.tuple.Pair;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-
 public class BlockItemPipe extends BlockMaterialPipe<ItemPipeType, ItemPipeProperties, WorldItemPipeNet> {
-
-    private final Map<Material, ItemPipeProperties> enabledMaterials = new HashMap<>();
 
     public BlockItemPipe(ItemPipeType itemPipeType, MaterialRegistry registry) {
         super(itemPipeType, registry);
         setCreativeTab(GTCreativeTabs.TAB_GREGTECH_PIPES);
         setHarvestLevel(ToolClasses.WRENCH, 1);
-    }
-
-    public void addPipeMaterial(Material material, ItemPipeProperties properties) {
-        Preconditions.checkNotNull(material, "material");
-        Preconditions.checkNotNull(properties, "material %s itemPipeProperties was null", material);
-        Preconditions.checkArgument(material.getRegistry().getNameForObject(material) != null,
-                "material %s is not registered", material);
-        this.enabledMaterials.put(material, properties);
     }
 
     @Override
@@ -67,11 +48,6 @@ public class BlockItemPipe extends BlockMaterialPipe<ItemPipeType, ItemPipePrope
     }
 
     @Override
-    protected ItemPipeProperties getFallbackType() {
-        return enabledMaterials.values().iterator().next();
-    }
-
-    @Override
     public WorldItemPipeNet getWorldPipeNet(World world) {
         return WorldItemPipeNet.getWorldPipeNet(world);
     }
@@ -82,27 +58,11 @@ public class BlockItemPipe extends BlockMaterialPipe<ItemPipeType, ItemPipePrope
         return ItemPipeRenderer.INSTANCE.getParticleTexture((TileEntityItemPipe) world.getTileEntity(blockPos));
     }
 
-    @Override
-    protected ItemPipeProperties createProperties(ItemPipeType itemPipeType, Material material) {
-        return itemPipeType.modifyProperties(enabledMaterials.getOrDefault(material, getFallbackType()));
-    }
-
     @SideOnly(Side.CLIENT)
     @NotNull
     @Override
     public PipeRenderer getPipeRenderer() {
         return ItemPipeRenderer.INSTANCE;
-    }
-
-    public Collection<Material> getEnabledMaterials() {
-        return Collections.unmodifiableSet(enabledMaterials.keySet());
-    }
-
-    @Override
-    public void getSubBlocks(@NotNull CreativeTabs itemIn, @NotNull NonNullList<ItemStack> items) {
-        for (Material material : enabledMaterials.keySet()) {
-            items.add(getItem(material));
-        }
     }
 
     @Override
